requirements.txt
Flask
numpy
opencv-python-headless
imageio
torch
torchvision
Pillow
gunicorn
timm
